+++ lib/active_record_compose/model.rb
@@ -28,7 +28,11 @@ module ActiveRecordCompose
       save_in_transaction { run_callbacks(:save) { save_models } }
     end
 
-    def save! = save || raise_on_save_error
+    def save!
+      valid? || raise_validation_error
+
+      save_in_transaction { run_callbacks(:save) { save_models } } || raise_on_save_error
+    end
 
     # Behavior is same to `#save`, but `before_create` and `after_create` hooks fires.
     #
@@ -65,7 +69,10 @@ module ActiveRecordCompose
     # Behavior is same to `#create`, but raises an exception prematurely on failure.
     #
     def create!(attributes = {})
-      create(attributes) || raise_on_save_error
+      assign_attributes(attributes)
+      valid? || raise_validation_error
+
+      save_in_transaction { run_callbacks(:save) { run_callbacks(:create) { save_models } } } || raise_on_save_error
     end
 
     # Behavior is same to `#save`, but `before_update` and `after_update` hooks fires.
@@ -103,7 +110,10 @@ module ActiveRecordCompose
     # Behavior is same to `#update`, but raises an exception prematurely on failure.
     #
     def update!(attributes = {})
-      update(attributes) || raise_on_save_error
+      assign_attributes(attributes)
+      valid? || raise_validation_error
+
+      save_in_transaction { run_callbacks(:save) { run_callbacks(:update) { save_models } } } || raise_on_save_error
     end
 
     private
@@ -129,6 +139,8 @@ module ActiveRecordCompose
     end
 
     def save_models = wrapped_models.all? { _1.save! }
+
+    def raise_validation_error = raise ActiveRecord::RecordInvalid, self
 
     def raise_on_save_error = raise ActiveRecord::RecordNotSaved.new('Failed to save the model.', self)
   end